% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpval.R
\name{bootpval}
\alias{bootpval}
\title{bootpval}
\usage{
bootpval(x, conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")
}
\arguments{
\item{x}{the sample}

\item{conf.level}{confidence level of the CI}

\item{iter}{iter times of the simulation}

\item{mu0}{the null hypothesis}

\item{test}{}
}
\value{

}
\description{
bootpval
}
