% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R, R/myboot3.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)

myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{...}{}
}
\value{
a histogram

a histogram
}
\description{
Takes a sample and using a confidence interval, makes a histogram with a point
estimate plotted 1/2 way up the density.

Takes a sample and using a confidence interval, makes a histogram with a point
estimate plotted 1/2 way up the density.
}
\examples{
set.seed(39); sam=rnorm(25,mean=25,sd=10)
myboot2(10000,x=sam,fun="mean",alpha=0.05,xlab="mean",col="Purple",cx=1.5) # mac quartz()



}
