% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{95% confidence interval function
Creates a 95% confidence interval for the mean from a single sample x}
\usage{
myci(x)
}
\arguments{
\item{x}{}
}
\value{
confidence interval with 95% confidence level
}
\description{
95% confidence interval function
Creates a 95% confidence interval for the mean from a single sample x
}
\examples{
x <- c(1,2,3,4,6,7,9)
}
