% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{CLT Binomial function}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{...}{}
}
\value{
a binominal density curve plot in the form of a histogram
}
\description{
Takes parameters n, iter, p and creates a histogram in the form of density plot.
n is the size, p is the probability, iter is the iterations
}
\examples{
mycltb(n=5,iter=10000,p=0.5)
}
