% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Poisson Distribution}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{Sample size}

\item{iter}{number of iterations}

\item{lambda}{lambda}

\item{...}{extra parameters}
}
\value{
histograms of sample mean, barplot of y, prob function for poisson
}
\description{
Takes a sample size, iteration, and a lambda and creates histograms of sample mean, barplot of sampled y, and plot of poisson probability function.
}
\examples{
mycltp(n=2,iter=10000,lambda=4)
}
