% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myconstr.R
\name{myconstr}
\alias{myconstr}
\title{Test Statistics, Mean and pvalue}
\usage{
myconstr(x, y, alpha)
}
\arguments{
\item{x}{data 1}

\item{y}{data 2}

\item{alpha}{confidence level}
}
\value{
an object
}
\description{
Test Statistics, Mean and pvalue
}
\examples{
\dont {obj <- mypackage::myconstr(x=x,y=y,alpha = 0.05)}
}
