% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Square Function
takes a vector and returns a vector of squared componets}
\usage{
myf(x)
}
\arguments{
\item{x}{}
}
\value{
a vector of swuared componets
}
\description{
Square Function
takes a vector and returns a vector of squared componets
}
\examples{
x=1:30; myf(x)
}
