% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function that x and param are evaluated based on}

\item{x}{vector of values}

\item{param}{another vector}

\item{...}{another param}
}
\value{
Maximum likelihood plot and other values
}
\description{
performs maximum likelihood estimate when given a function, a vector of values and a vector param.
}
