% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal distribution curve}
\usage{
myncurve(mu, sigma, p)
}
\arguments{
\item{mu}{mean of the distribution}

\item{sigma}{standard deviation of the distribution}

\item{p}{start point to calculate the lower tail}
}
\value{
returns a graph of the distribution and the area of the lower tail probability, also gives you the lower tail probability
}
\description{
Normal distribution curve
}
