% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{mypvalue function
this function creates a plot of the p value given certain parameters below.}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{this is your vertical line quantile value, (t alpha/2)}

\item{xmax}{this is the length of your x axis}

\item{n}{n is your sample size}

\item{alpha}{this is your default alpha value}
}
\value{
a plot of the p value with the interval and shaded region (acceptance and rejection regions.).
}
\description{
mypvalue function
this function creates a plot of the p value given certain parameters below.
}
\examples{
mypvalue(tcalc,n=30,alpha=0.05)
t.test(x1,mu=23)

}
