% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{nticket function}
\usage{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{Number of seats available on the plane}

\item{gamma}{probability that a plan will be truly overbooked}

\item{p}{probability that an individual will show up for their flight}
}
\value{
2 plots and a list containing nd, nc, N, p, and gamma
}
\description{
This function computes and plots the ideal number of tickets to sell to optimize the number of seats filled based on number of tickets sold. 2 plots are made, one depicting a discrete distribution, and one depicting a continuous distribution.
}
\examples{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
