% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{pareto}
\usage{
pareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{x}{}
}
\value{
Creates a pareto barplot using a replicated vector form by frequencies with their respective names
}
\description{
pareto
}
\examples{
freq=c(15,8,63,20) # Vector of frequencies of different types of robot limbs
names=c("None","Both","LegsO","WheelsO") # Respective names of the robot libmbs according to their frequencies
l=rep(names,freq) # Replicated vector
pareto(l) #Pareto diagram output

}
