% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Rttest.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{Plots a boxplot side by side of x and y}
\usage{
\method{plot}{Rttest}(obj, ...)
}
\arguments{
\item{obj}{- This is the object of class Rttest}

\item{...}{}
}
\value{
returns a box plot
}
\description{
Plots a boxplot side by side of x and y
}
\examples{
\dontrun{plot(obj)}
}
