% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Rttest.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Title}
\usage{
\method{print}{Rttest}(obj, ...)
}
\arguments{
\item{obj}{-object}
}
\value{
a table with the data and the alpha value, confidentervaks and p-value
}
\description{
Title
}
\examples{
\donotrun print(obj)
}
