% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{quadfun}
\alias{quadfun}
\title{Square Function}
\usage{
quadfun(x)
}
\arguments{
\item{x}{}
}
\value{
output in the object quad.lm.
}
\description{
this function works by placing the output in the object quad.lm by making it a function of x
}
