% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Simulation}
\usage{
mybin(iter = 1000, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{Number of samples taken}

\item{p}{Probability of success}
}
\value{
plot and such
}
\description{
Binomial Simulation
}
\examples{
mybin()

}
