% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{My boot function (bootstrap simulation technique)}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{Number of iterations of the simulation}

\item{x}{population sample}

\item{fun}{function for the bootstrap intervals (ex mean, variance, etc)}

\item{alpha}{alpha level (1-alpha is equal to the confidence interval)}
}
\value{
Returns a histogram of the sample statistic which also shows the confidence interval for the mean at the established 1-alpha confidence level
}
\description{
My boot function (bootstrap simulation technique)
}
\examples{
For a basic sample s=c(3,4,5,6) use myboot(x=s) and you will get a 95\% confidence interval for the mean
}
