% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of iterations}

\item{x}{sample}

\item{fun}{function to do to x}

\item{alpha}{confidence interval}

\item{cx}{not sure}

\item{...}{dot dot dot}
}
\value{
charts
}
\description{
performs a bootstrap and returns charts
}
