% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{myci function (Confidence interval for the mean of a single sample)}
\usage{
myci(x, ci = 0.95, ...)
}
\arguments{
\item{x}{sample}

\item{ci}{desired confidence interval}

\item{...}{}
}
\value{
returns a range of the desired confidence interval
}
\description{
myci function (Confidence interval for the mean of a single sample)
}
