% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Square function}
\usage{
myf(x)
}
\arguments{
\item{x}{}
}
\value{
a vector of squared components
}
\description{
takes vector and returns vector of squared compnents
}
\examples{
x=1:20; myf(x)
}
