% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Square function}
\usage{
myf(x)
}
\arguments{
\item{x}{}
}
\value{
a Vector of squared components
}
\description{
Takes a vector and returns a vector of squared components
}
\details{
The function is an elementary example of function creation and can be massively enhanced
}
\examples{
x=1:30; myf(x)
}
