% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Square Function}
\usage{
myf(x)
}
\arguments{
\item{x}{is the user input}
}
\value{
a vector of squared components in the form of a vector object.
}
\description{
Takes a vector and returns a vector of squared components.
}
\examples{
myf(x = 1:30)

}
