% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Probability Distribution}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{probability}
}
\value{
Normal plot with shaded area probability
}
\description{
This function takes an input mean, standard deviation, and an a value and makes a normal curve with the lower tail area below a shaded in. The area of this shaded region is determined and displayed.
}
\examples{
\dontrun{myncurve(mu = 3, sigma = 5, a = 6)}

}
