% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Rttest.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Prints the x, y using the kable package}
\usage{
\method{print}{Rttest}(obj, ...)
}
\arguments{
\item{obj}{-object}
}
\value{
a table with the data and the alpha value, confidentervaks and p-value
}
\description{
Prints the x, y using the kable package
}
\examples{
\dontrun{
print(obj)}
}
